Created a controlled input for the location and Added changes for Counter component
