+++ src/App.js
@@ -1,40 +1,34 @@
 import React from "react";
 
-class Counter extends React.Component {
+class App extends React.Component {
   constructor(props) {
     super(props);
-
-    this.state = { counter: 0 };
-    this.handleIncrement = this.handleIncrement.bind(this);
-    this.handleDecrement = this.handleDecrement.bind(this);
+    this.state = { location: "Lisbon" };
+    this.fetchWeather = this.fetchWeather.bind(this);
   }
-
-  handleIncrement() {
-    // console.log("handleAdd triggered");
-    // console.log("this", this);
-    this.setState((curState) => {
-      return { counter: curState.counter + 1 };
-    });
-  }
-
-  handleDecrement() {
-    if (this.state.counter === 0) return;
-    this.setState((curState) => {
-      return {
-        counter: curState.counter - 1,
-      };
-    });
+  fetchWeather() {
+    console.log("loading data...");
+    console.log("this", this);
   }
 
   render() {
     return (
-      <div>
-        <button onClick={this.handleIncrement}>+</button>
-        <span>{this.state.counter}</span>
-        <button onClick={this.handleDecrement}>-</button>
+      <div className="app">
+        <h1>Classy-weather</h1>
+        <div className="input">
+          <input
+            type="text"
+            placeholder="Search for location"
+            value={this.state.location}
+            onChange={(e) => this.setState({ location: e.target.value })}
+          />
+        </div>
+        <button className="input" onClick={this.fetchWeather}>
+          Get weather
+        </button>
       </div>
     );
   }
 }
 
-export default Counter;
+export default App;
